Renamed storage to file_storage
